ENH: moved filter and added axis arg      moved where,mask,align
TST: make reindex benchmarks longer

CLN: fixed up names for creation in panelnd.py

DOC: minor release notes changes
