+++ pandas/core/generic.py
@@ -177,6 +177,8 @@ class PandasObject(object):
 
     def _get_block_manager_axis(self, axis):
         """ map the axis to the block_manager axis """
+        if axis not in self._AXIS_NAMES:
+            raise Exception("this object does not support an axis of [%s]" % axis)
         if self._AXIS_REVERSED:
             m = self._AXIS_LEN-1
             return m-axis
@@ -559,13 +561,14 @@ class PandasObject(object):
         -------
         selection : type of caller
         """
-        axis_name = self._get_axis_name(axis)
-        axis = self._get_axis(axis)
+        axis        = self._get_axis_number(axis)       
+        axis_name   = self._get_axis_name(axis)
+        axis_values = self._get_axis(axis)
 
-        if len(axis) > 0:
-            new_axis = axis[np.asarray([bool(crit(label)) for label in axis])]
+        if len(axis_values) > 0:
+            new_axis = axis_values[np.asarray([bool(crit(label)) for label in axis_values])]
         else:
-            new_axis = axis
+            new_axis = axis_values
 
         return self.reindex(**{axis_name: new_axis})
 
@@ -696,6 +699,40 @@ class PandasObject(object):
             raise ValueError(msg)
 
         return self.shift(periods, freq, **kwds)
+
+    def truncate(self, before=None, after=None, copy=True):
+        """Function truncate a sorted DataFrame / Series before and/or after
+        some particular dates.
+
+        Parameters
+        ----------
+        before : date
+        Truncate before date
+        after : date
+        Truncate after date
+
+        Returns
+        -------
+        truncated : type of caller
+        """
+        from pandas.tseries.tools import to_datetime
+        before = to_datetime(before)
+        after = to_datetime(after)
+        
+        if before is not None and after is not None:
+            if before > after:
+                raise AssertionError('Truncate: %s must be after %s' %
+                                     (before, after))
+
+        result = self.ix[before:after]
+
+        if isinstance(self.index, MultiIndex):
+            result.index = self.index.truncate(before, after)
+
+        if copy:
+            result = result.copy()
+
+        return result
 
     def pct_change(self, periods=1, fill_method='pad', limit=None, freq=None,
                    **kwds):
@@ -1188,6 +1225,46 @@ class NDFrame(PandasObject):
         else:
             return self._constructor(new_data)
 
+    def filter(self, items=None, like=None, regex=None, axis=None):
+        """
+        Restrict the info axis to set of items or wildcard
+
+        Parameters
+        ----------
+        items : list-like
+            List of info axis to restrict to (must not all be present)
+        like : string
+            Keep info axis where "arg in col == True"
+        regex : string (regular expression)
+            Keep info axis with re.search(regex, col) == True
+
+        Notes
+        -----
+        Arguments are mutually exclusive, but this is not checked for
+
+        Returns
+        -------
+        same type as input object with filtered info axis
+        """
+        import re
+
+        if axis is None:
+            axis = self._info_axis_name
+        axis_name   = self._get_axis_name(axis)
+        axis_values = self._get_axis(axis_name)
+
+        if items is not None:
+            return self.reindex(**{ axis_name : [r for r in items if r in axis_values ] })
+        elif like:
+            matchf = lambda x: (like in x if isinstance(x, basestring)
+                                else like in str(x))
+            return self.select(matchf, axis=axis_name)
+        elif regex:
+            matcher = re.compile(regex)
+            return self.select(lambda x: matcher.search(x) is not None, axis=axis_name)
+        else:
+            raise ValueError('items was None!')
+
     #----------------------------------------------------------------------
     # Attribute access
 
@@ -1346,7 +1423,7 @@ class NDFrame(PandasObject):
         mgr = self._data.astype(dtype, copy = copy, raise_on_error = raise_on_error)
         return self._constructor(mgr)
 
-    def convert_objects(self, convert_dates=True, convert_numeric=True):
+    def convert_objects(self, convert_dates=True, convert_numeric=False):
         """
         Attempt to infer better dtype for object columns
         Always returns a copy (even if no object columns)
@@ -1358,9 +1435,211 @@ class NDFrame(PandasObject):
 
         Returns
         -------
-        converted : DataFrame
+        converted : asm as input object
         """
         return self._constructor(self._data.convert(convert_dates=convert_dates, convert_numeric=convert_numeric))
+
+    def align(self, other, join='outer', axis=None, level=None, copy=True,
+              fill_value=np.nan, method=None, limit=None, fill_axis=0):
+        """
+        Align two object on their axes with the
+        specified join method for each axis Index
+
+        Parameters
+        ----------
+        other : DataFrame or Series
+        join : {'outer', 'inner', 'left', 'right'}, default 'outer'
+        axis : allowed axis of the other object, default None
+            Align on index (0), columns (1), or both (None)
+        level : int or name
+            Broadcast across a level, matching Index values on the
+            passed MultiIndex level
+        copy : boolean, default True
+            Always returns new objects. If copy=False and no reindexing is
+            required then original objects are returned.
+        fill_value : scalar, default np.NaN
+            Value to use for missing values. Defaults to NaN, but can be any
+            "compatible" value
+        method : str, default None
+        limit : int, default None
+        fill_axis : {0, 1}, default 0
+            Filling axis, method and limit
+
+        Returns
+        -------
+        (left, right) : (type of input, type of other)
+            Aligned objects
+        """
+        from pandas import DataFrame,Series
+
+        if isinstance(other, DataFrame):
+            return self._align_frame(other, join=join, axis=axis, level=level,
+                                     copy=copy, fill_value=fill_value,
+                                     method=method, limit=limit,
+                                     fill_axis=fill_axis)
+        elif isinstance(other, Series):
+            return self._align_series(other, join=join, axis=axis, level=level,
+                                      copy=copy, fill_value=fill_value,
+                                      method=method, limit=limit,
+                                      fill_axis=fill_axis)
+        else:  # pragma: no cover
+            raise TypeError('unsupported type: %s' % type(other))
+
+    def _align_frame(self, other, join='outer', axis=None, level=None,
+                     copy=True, fill_value=np.nan, method=None, limit=None,
+                     fill_axis=0):
+        # defaults
+        join_index, join_columns = None, None
+        ilidx, iridx = None, None
+        clidx, cridx = None, None
+
+        if axis is None or axis == 0:
+            if not self.index.equals(other.index):
+                join_index, ilidx, iridx = \
+                    self.index.join(other.index, how=join, level=level,
+                                    return_indexers=True)
+
+        if axis is None or axis == 1:
+            if not self.columns.equals(other.columns):
+                join_columns, clidx, cridx = \
+                    self.columns.join(other.columns, how=join, level=level,
+                                      return_indexers=True)
+
+        left  = self._reindex_with_indexers({ 0 : [ join_index,   ilidx ],
+                                              1 : [ join_columns, clidx ] },
+                                            copy=copy, fill_value=fill_value)
+        right = other._reindex_with_indexers({ 0 : [ join_index,   iridx ],
+                                               1 : [ join_columns, cridx ] }, 
+                                             copy=copy, fill_value=fill_value)
+            
+
+        if method is not None:
+            left = left.fillna(axis=fill_axis, method=method, limit=limit)
+            right = right.fillna(axis=fill_axis, method=method, limit=limit)
+
+        return left, right
+
+    def _align_series(self, other, join='outer', axis=None, level=None,
+                      copy=True, fill_value=None, method=None, limit=None,
+                      fill_axis=0):
+        from pandas import DataFrame
+
+        fdata = self._data
+        if axis == 0:
+            join_index = self.index
+            lidx, ridx = None, None
+            if not self.index.equals(other.index):
+                join_index, lidx, ridx = self.index.join(other.index, how=join,
+                                                         return_indexers=True)
+
+            if lidx is not None:
+                fdata = fdata.reindex_indexer(join_index, lidx, axis=1)
+        elif axis == 1:
+            join_index = self.columns
+            lidx, ridx = None, None
+            if not self.columns.equals(other.index):
+                join_index, lidx, ridx = \
+                    self.columns.join(other.index, how=join,
+                                      return_indexers=True)
+
+            if lidx is not None:
+                fdata = fdata.reindex_indexer(join_index, lidx, axis=0)
+        else:
+            raise ValueError('Must specify axis=0 or 1')
+
+        if copy and fdata is self._data:
+            fdata = fdata.copy()
+
+        left_result = DataFrame(fdata)
+        right_result = other if ridx is None else other.reindex(join_index)
+
+        fill_na = notnull(fill_value) or (method is not None)
+        if fill_na:
+            return (left_result.fillna(fill_value, method=method, limit=limit,
+                                       axis=fill_axis),
+                    right_result.fillna(fill_value, method=method,
+                                        limit=limit))
+        else:
+            return left_result, right_result
+
+    def where(self, cond, other=np.nan, inplace=False, try_cast=False, raise_on_error=True):
+        """
+        Return an object of same shape as self and whose corresponding
+        entries are from self where cond is True and otherwise are from other.
+
+        Parameters
+        ----------
+        cond : boolean DataFrame or array
+        other : scalar or DataFrame
+        inplace : boolean, default False
+            Whether to perform the operation in place on the data
+        try_cast : boolean, default False
+            try to cast the result back to the input type (if possible),
+        raise_on_error : boolean, default True
+            Whether to raise on invalid data types (e.g. trying to where on
+            strings)
+
+        Returns
+        -------
+        wh : DataFrame
+        """
+        if isinstance(cond, NDFrame):
+            cond = cond.reindex(**self._construct_axes_dict())
+        else:
+            if not hasattr(cond, 'shape'):
+                raise ValueError('where requires an ndarray like object for its '
+                                 'condition')
+            if cond.shape != self.shape:
+                raise ValueError('Array conditional must be same shape as self')
+            cond = self._constructor(cond, index=self.index,
+                                     columns=self.columns)
+
+        if inplace:
+            cond = -(cond.fillna(True).astype(bool))
+        else:
+            cond = cond.fillna(False).astype(bool)
+
+        # try to align
+        if hasattr(other, 'align'):
+
+            # align with me
+            if other.ndim <= self.ndim:
+
+                _, other = self.align(other, join='left', fill_value=np.nan)
+
+            # slice me out of the other
+            else:
+                raise NotImplemented
+            
+        elif isinstance(other,np.ndarray):
+            if other.shape != self.shape:
+                raise ValueError('other must be the same shape as self '
+                                 'when an ndarray')
+            other = self._constructor(other, **self._construct_axes_dict())
+
+        if inplace:
+            # we may have different type blocks come out of putmask, so reconstruct the block manager
+            self._data = self._data.putmask(cond,other,inplace=True)
+
+        else:
+            new_data = self._data.where(other, cond, raise_on_error=raise_on_error, try_cast=try_cast)
+
+            return self._constructor(new_data)
+
+    def mask(self, cond):
+        """
+        Returns copy of self whose values are replaced with nan if the
+        inverted condition is True
+
+        Parameters
+        ----------
+        cond: boolean object or array
+
+        Returns
+        -------
+        wh: same as input
+        """
+        return self.where(~cond, np.nan)
 
     def cumsum(self, axis=None, skipna=True):
         """
@@ -1653,39 +1932,3 @@ class NDFrame(PandasObject):
 
         return new_obj
 
-# Good for either Series or DataFrame
-
-
-def truncate(self, before=None, after=None, copy=True):
-    """Function truncate a sorted DataFrame / Series before and/or after
-    some particular dates.
-
-    Parameters
-    ----------
-    before : date
-        Truncate before date
-    after : date
-        Truncate after date
-
-    Returns
-    -------
-    truncated : type of caller
-    """
-    from pandas.tseries.tools import to_datetime
-    before = to_datetime(before)
-    after = to_datetime(after)
-
-    if before is not None and after is not None:
-        if before > after:
-            raise AssertionError('Truncate: %s must be after %s' %
-                                 (before, after))
-
-    result = self.ix[before:after]
-
-    if isinstance(self.index, MultiIndex):
-        result.index = self.index.truncate(before, after)
-
-    if copy:
-        result = result.copy()
-
-    return result
